7 nov progressbar, markers on map (locally)

+++ app/src/main/java/com/l2minigames/wanderfulworld/MyMarker.java
@@ -4,15 +4,15 @@ package com.l2minigames.wanderfulworld;
  * Created by umyhlarsle on 2016-11-03.
  */
 public class MyMarker {
-    long markerLatitude;
-    long markerLongitude;
+    double markerLatitude;
+    double markerLongitude;
     String markerType;
 
     public MyMarker(){
 
     }
 
-    public MyMarker(long markerLatitude, long markerLongitude, String markerType){
+    public MyMarker(double markerLatitude, double markerLongitude, String markerType){
 
         this.markerLatitude = markerLatitude;
         this.markerLongitude = markerLongitude;
